redesigning app structure
initially setup project seperating styled components from their JSX component. Which made it hard to find which styles affect what component. Now coupling styles and components first by use then by component name
