+++ src/components/singleProduct/makeOffer/offerStyles/backCardStyles.js
