+++ src/components/singleProduct/makeOffer/offerStyles/frontCardStyles.js
