+++ src/components/singleProduct/makeOffer/offerStyles/makeOfferStyles.js
